script

+++ Assets/Scripts/OpenDoor.cs
@@ -15,7 +15,7 @@ public class OpenDoor : MonoBehaviour
     public float sensorStrength = 1;
 
     float distPacman = 0f, distBlinky = 0f, distInky = 0f, distPinky = 0f, distClyde =0f;
-    float door_open_num = 0f;
+    int door_open_num = 0;
 
     // Start is called before the first frame update
     void Start()
@@ -57,7 +57,7 @@ public class OpenDoor : MonoBehaviour
             Color tempDoorColor = doorRenderer.color;
             tempDoorColor.a = 0f;
             doorRenderer.color = tempDoorColor;
-            door_open_num += 1f;
+            door_open_num += 1;
         }
         else
         {
@@ -65,5 +65,6 @@ public class OpenDoor : MonoBehaviour
             tempDoorColor.a = 1f;
             doorRenderer.color = tempDoorColor;
         }
+
     }
 }
